Improve event handling in c++ example code
The c++ example application should show the basics of using
the library from C++, but should not validate the exact 
behavior of it, as that's the role of the unit tests.

This change refactors the example application to be more robust
to in which order events are generated. It also fixes a bug where
any event that was not understood was seen as "no event".