Restructure CXX example code
The example now live in a directory that explains what it does.

The include/dcsctp.h src/lib.cpp files have been removed since they
wasn't strictly needed and only made the build pipeline more
complicated.

The Makefile is now more idiomatic and better show how each file is
generated and what it depends on. Also, the run step was removed since
that's not traditionally part of a Makefile.